# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FollowUpBussiness API
  version: 1.0.0
  summary: Contrato REST general del MVP
  description: |
    API multiempresa para administración, aplicación móvil y procesos de
    integración de FollowUpBussiness.

    Reglas globales:
    - El tenant y el usuario se derivan de la identidad autenticada.
    - Ningún `tenantId` recibido en payload se considera autoridad.
    - Las colecciones usan paginación por página y orden estable.
    - Fechas y horas se expresan en ISO 8601; la fecha operativa usa la zona
      horaria de la empresa.
    - Importes usan decimal; el servidor recalcula totales.
    - Comandos mobile sensibles requieren idempotencia.
    - Toda operación REST acepta opcionalmente `X-Correlation-Id` conforme a
      `correlation-id-contract.md`, y devuelve el ID efectivo tanto en éxito
      como en error.
    - Auth WEB usa refresh solo en cookie HttpOnly; MOBILE usa body solo fuera
      de contexto navegador. El canal nunca concede identidad, tenant o rol.
  contact:
    name: FollowUpBussiness Engineering
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
  x-contract-status: READY_FOR_HANDOFF
  x-source-of-truth:
    - ../../00_CONTRATO_FUNCIONAL.md
    - ../stories/enablers/EN-013-definir-autenticacion-sesiones-y-recuperacion.md
    - ../architecture/adr/ADR-008-autenticacion-sesiones.md
    - ../stories/BACKLOG_REFINEMENT.md
    - ../stories/dependency-map.md
servers:
  - url: /api/v1
...
    variables:
...
security:
  - bearerAuth: []
...
paths:
...
  /routes:
    get:
      tags: [Routes]
      operationId: listRoutes
      summary: Listar rutas del tenant
      x-required-roles: [COMPANY_ADMIN, SUPERVISOR]
      x-story-ids: [BE-061, FE-014]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: date
          in: query
          schema: { type: string, format: date }
        - name: sellerId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/RouteStatus' }
      responses:
        '200':
          description: Página de rutas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoutePage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Routes]
      operationId: createRoute
      summary: Crear una ruta manual
      x-required-roles: [COMPANY_ADMIN, SUPERVISOR]
      x-story-ids: [BE-021, FE-015, INT-007]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRouteRequest' }
      responses:
        '201':
          description: Ruta creada
          headers:
            Location:
              schema: { type: string, format: uri-reference }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Route' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /routes/{routeId}:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    get:
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      tags: [Routes]
      operationId: getRoute
      summary: Consultar una ruta
      x-required-roles: [COMPANY_ADMIN, SUPERVISOR, SELLER]
      x-story-ids: [BE-061, FE-014, MOB-004]
      responses:
        '400': { $ref: '#/components/responses/CorrelationIdInvalid' }
        '200':
          description: Ruta
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Route' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Routes]
      operationId: updateRoute
      summary: Editar una ruta en borrador
      x-required-roles: [COMPANY_ADMIN, SUPERVISOR]
      x-story-ids: [BE-021, FE-015, INT-007]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/UpdateRouteRequest' }
      responses:
        '200':
          description: Ruta actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Route' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
  /routes/{routeId}/directions:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    get:
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      tags: [Routes]
      operationId: getRouteDirections
      summary: Obtener detalle vial efímero de una ruta
      description: El proveedor recibe exclusivamente coordenadas WGS84; usa el inicio explícito cuando existe o el primer punto ordenado como inicio, y el último punto como destino. Sin inicio explícito se requieren al menos dos puntos ubicados. El resultado no se persiste.
      x-required-roles: [COMPANY_ADMIN, SUPERVISOR, SELLER]
      responses:
        '200':
          description: Geometría, tramos e indicaciones viales neutrales
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteDirectionsResponse' }
        '400': { $ref: '#/components/responses/CorrelationIdInvalid' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '503': { $ref: '#/components/responses/RouteDirectionsUnavailable' }
  /routes/{routeId}/directions/preview:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    post:
      tags: [Routes]
      operationId: previewRouteDirections
      summary: Previsualizar detalle vial para un orden propuesto
      description: Calcula de forma efímera la geometría vial de una permutación completa de puntos de un borrador. No modifica la ruta ni persiste el resultado.
      x-required-roles: [COMPANY_ADMIN, SUPERVISOR]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PreviewRouteDirectionsRequest' }
      responses:
        '200':
          description: Geometría, tramos e indicaciones viales neutrales para el orden propuesto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteDirectionsResponse' }
        '400': { $ref: '#/components/responses/CorrelationIdInvalid' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '503': { $ref: '#/components/responses/RouteDirectionsUnavailable' }
  /routes/{routeId}/points/order:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    put:
      tags: [Routes]
      operationId: reorderRoutePoints
      summary: Reordenar todos los puntos de una ruta DRAFT o PUBLISHED antes de iniciar la jornada
      x-required-roles: [COMPANY_ADMIN, SUPERVISOR]
      x-story-ids: [BE-023, BE-064, FE-015, FE-016]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReorderRoutePointsRequest' }
      responses:
        '200':
          description: Ruta reordenada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Route' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/RoutePlanningSnapshotConflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /routes/{routeId}/publish:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    post:
      tags: [Routes]
      operationId: publishRoute
      summary: Publicar ruta y disparar su notificación
      x-required-roles: [COMPANY_ADMIN, SUPERVISOR]
      x-story-ids: [BE-024, BE-053, FE-017, MOB-029, INT-007]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishRouteRequest' }
      responses:
        '400': { $ref: '#/components/responses/CorrelationIdInvalid' }
        '200':
          description: Ruta publicada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Route' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /routes/{routeId}/reassign:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    post:
      tags: [Routes]
      operationId: reassignRoute
      summary: Reasignar una ruta
      description: >-
        COMPANY_ADMIN puede reasignar rutas de su tenant. SUPERVISOR solo puede
        reasignar cuando el vendedor actual y el nuevo pertenecen a su equipo
        vigente. Solo admite PUBLISHED; las visitas completadas conservan
        responsable histórico y los puntos
        pendientes habilitan al nuevo vendedor solo para esta ruta, sin alterar
        cartera ni jornadas.
      x-required-roles: [COMPANY_ADMIN, SUPERVISOR]
      x-story-ids: [BE-025, FE-018, INT-009]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReassignRouteRequest' }
      responses:
        '400': { $ref: '#/components/responses/CorrelationIdInvalid' }
        '200':
          description: Ruta reasignada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Route' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /routes/{routeId}/copy:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    post:
      tags: [Routes]
      operationId: copyRoute
      summary: Duplicar una ruta para otra fecha
      x-required-roles: [COMPANY_ADMIN, SUPERVISOR]
      x-story-ids: [BE-026, FE-015]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CopyRouteRequest' }
      responses:
        '400': { $ref: '#/components/responses/CorrelationIdInvalid' }
        '201':
          description: Copia en borrador
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CopyRouteResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /routes/suggested-customers:
    get:
      tags: [Routes]
      operationId: listSuggestedCustomers
      summary: Sugerir clientes por frecuencia de visita
      x-required-roles: [COMPANY_ADMIN, SUPERVISOR]
      x-story-ids: [BE-027, FE-016]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: sellerId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Clientes sugeridos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuggestedCustomerPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT RS256 de acceso, 10 minutos, transportado solo en Authorization; nunca en cookie.
    webRefreshCookie:
      type: apiKey
      in: cookie
      name: __Host-fs-refresh
      description: Cookie refresh exclusiva de WEB; Secure, HttpOnly, SameSite=Strict y Path=/.
    sessionRevocationTicket:
      type: apiKey
      in: header
      name: X-Session-Revocation-Ticket
      description: Ticket opaco MOBILE de un uso y solo para revocar su propia familia; no autentica ni permite refresh.
...
  parameters:
...
    Page:
      name: page
      in: query
      description: Página basada en cero.
      schema: { type: integer, minimum: 0, default: 0 }
    PageSize:
      name: pageSize
      in: query
      description: Cantidad máxima de elementos por página.
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
...